backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
langchain
chromadb
sentence-transformers
reportlab
python-docx